--- packages/graphql-playground-react/src/reducers/index.ts
@@ -1,8 +1,0 @@
-import { combineReducers, Reducer } from 'redux'
-import docs from './docs'
-
-const combinedReducers: Reducer<any> = combineReducers({
-  docs,
-})
-
-export default combinedReducers
