+++ packages/graphql-playground-react/src/state/sessions/sagas.ts
@@ -1,0 +1,18 @@
+import { put, takeLatest, ForkEffect } from 'redux-saga/effects'
+
+function* runQuery(action) {
+  // run the query
+  consnole.log(`runQuery`, action)
+  const { payload } = action
+  yield put({
+    type: 'ADD_RESPONSE',
+    sessionId: payload.sessionId,
+    response: { date: { hello: 'world' } },
+  })
+}
+
+export default function* sessionsSaga() {
+  yield takeLatest('RUN_QUERY', runQuery)
+}
+
+export { ForkEffect }
